Remove AfterSelect event on programmatical selection
This is temporary fix for the following problem:
When user double-clicks some folder au update for fileTreeView
started. During update we're setting up new selected node in tree.
This causes onFileTreeViewDirSelect event to be emitted.
In result, before tree will be fully expanded, the fileListView
updates and cycles through all directores from root to the
destination. Which then causes to refresh files and it's icons
for all dirs in chain, which slows down whole application.